+++ R/methods-MSnExp.R
@@ -450,7 +450,7 @@ setMethod("splitByFile", c("MSnExp", "factor"), function(object, f) {
 #'     depend on the architecture. Default is
 #'     \code{BiocParallel::bpparam()}.
 #'
-#' @return \code{chromatogram} returns a \code{\link{Chromatograms}} object with
+#' @return \code{chromatogram} returns a \code{\link{MChromatograms}} object with
 #'     the number of columns corresponding to the number of files in
 #'     \code{object} and number of rows the number of specified ranges (i.e.
 #'     number of rows of matrices provided with arguments \code{mz} and/or
@@ -461,7 +461,7 @@ setMethod("splitByFile", c("MSnExp", "factor"), function(object, f) {
 #'
 #' @author Johannes Rainer
 #'
-#' @seealso \code{\link{Chromatogram}} and \code{\link{Chromatograms}} for the
+#' @seealso \code{\link{Chromatogram}} and \code{\link{MChromatograms}} for the
 #'     classes that represent single and multiple chromatograms.
 #'
 #' @examples
@@ -491,7 +491,7 @@ setMethod("splitByFile", c("MSnExp", "factor"), function(object, f) {
 #' mzr <- rbind(c(140, 160), c(300, 320))
 #' chrs <- chromatogram(msd, rt = rtr, mz = mzr)
 #'
-#' ## Each row of the returned Chromatograms object corresponds to one mz-rt
+#' ## Each row of the returned MChromatograms object corresponds to one mz-rt
 #' ## range. The Chromatogram for the first range in the first file is empty,
 #' ## because the retention time range is outside of the file's rt range:
 #' chrs[1, 1]
@@ -526,7 +526,7 @@ setMethod("chromatogram", "MSnExp", function(object, rt, mz,
                                          missingValue = missing,
                                          msLevel = msLevel,
                                          BPPARAM = BPPARAM)
-    res <- as(res, "Chromatograms")
+    res <- as(res, "MChromatograms")
     if (!nrow(res))
         return(res)
     fd <- annotatedDataFrameFrom(res, byrow = TRUE)